Making start/end times and dates optional
The system only allowed for events with start and end times to be added
previously.  Now we also allow for events with no start/end times, or
events with only start times.

Note: There are extra indentation and whitespace cleanups in this commit
due to my editor cleaning them up by default.  Apologies for the clutter.