Adding curate menu & allowing greeting editing
The original welcome page content required manual editing in the
template, but CityHapps wanted that content to be more readily available
for editing.  As such its content has been brought into the database and
the ability to edit it has been added to a new "Curate" section in the
menu.

The migration for this change includes the import statement to bring the
original content into the DB so it will be ready to go in Production,
QA, or other dev environments.