Fixing admin routing authentication
The prior default behavior was set to try to authenticate against a
controller method in a way that would never succeed unless accessed
through the angular router (so a normal HTTP request could never start
there).  This was incredibly inconvenient and ultimately did not protect
the page in the intended way.

I've setup the angular router to authenticate and authorize all
necessary routes using the angular router's existing conventions.  If
someone unauthenticated tries to access an authenticated route, the
client will redirect them to the login page. If they are authenticated
they may proceed as normal.  If someone attempts to access an admin
route, they must be an admin or else they will be redirected to the home
page.